requirements.txt
certifi==2018.10.15
chainer==5.0.0
chardet==3.0.4
fastrlock==0.4
filelock==3.0.10
future==0.17.1
gym==0.10.9
idna==2.7
numpy==1.15.4
Pillow==5.3.0
protobuf==3.6.1
pyglet==1.3.2
requests==2.20.1
scipy==1.1.0
six==1.11.0
urllib3==1.24.1
